fix(imaging): gate prepareImage fallback on WebP magic + cap error; add ErrResolutionCapExceeded
Two issues flagged in review:

1. decode_test.go line 151: `size` was int, matching the old production
   bug we already fixed. Change to int64 for consistency with decode.go.

2. prepareImage fallback was unconditional — it retried any
   DecodeMemBounded failure with DecodeWebPFirstFrame, including
   resolution cap rejections and non-WebP files.

Fix: export ErrResolutionCapExceeded sentinel from the imaging package
(wrapped into checkConfigResolution via %w), then in prepareImage:
- Skip the fallback entirely if errors.Is detects a cap rejection.
  A cap-rejected image must never bypass the resolution limit through
  a second decoder path.
- Also guard with a 12-byte RIFF/WEBP magic check before the retry so
  non-WebP files skip the seek and the DecodeWebPFirstFrame call
  entirely.

+++ server/channels/app/imaging/decode_test.go
@@ -148,12 +148,12 @@ func TestDecodeWebPFirstFrame(t *testing.T) {
 	var vp8Chunk []byte
 	for off := 12; off+8 <= len(staticData); {
 		id := string(staticData[off : off+4])
-		size := int(binary.LittleEndian.Uint32(staticData[off+4 : off+8]))
+		size := int64(binary.LittleEndian.Uint32(staticData[off+4 : off+8]))
 		if id == "VP8 " {
-			vp8Chunk = staticData[off : off+8+size]
+			vp8Chunk = staticData[off : off+8+int(size)]
 			break
 		}
-		off += 8 + size + (size % 2)
+		off += 8 + int(size) + int(size%2)
 	}
 	require.NotEmpty(t, vp8Chunk, "VP8 chunk not found in testwebp.webp")
